requirements.txt
# Main Project Dependencies
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.0
requests==2.31.0
Pillow==10.1.0
twilio==8.10.0

# Frontend
streamlit==1.28.0
streamlit-mic-recorder==0.0.5
